Created Api's to Create and mark resolved the Grievances by the user, and reflect the same on Admin Dashboard and Faculty Dashboard

+++ src/main/java/com/example/grievancecell/repository/UserRepository.java
@@ -1,10 +1,8 @@
 package com.example.grievancecell.repository;
 
-
 import com.example.grievancecell.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
-
